Volley related everything removed.
