+++ app/build.gradle
@@ -146,7 +146,6 @@ dependencies {
 
     //Google
     compile 'com.google.code.gson:gson:2.3.1'
-    compile 'com.google:volley:4.3'
     compile 'com.google.http-client:google-http-client-gson:1.19.0'
     compile 'com.google.apis:google-api-services-plus:v1-rev126-1.18.0-rc'
     compile 'com.google.android.apps.dashclock:dashclock-api:2.0.0'
