--- app/src/main/java/org/gdg/frisbee/android/api/ApiException.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2013-2015 The GDG Frisbee Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gdg.frisbee.android.api;
-
-/**
- * GDG Aachen
- * org.gdg.frisbee.android.api
- * <p/>
- * User: maui
- * Date: 21.04.13
- * Time: 22:24
- */
-public class ApiException extends Exception {
-    public ApiException(Throwable e) {
-        super(e);
-    }
-
-    public ApiException(String msg) {
-        super(msg);
-    }
-
-    public ApiException(String msg, int code) {
-        super(code + " " + msg);
-    }
-}
